+++ src/components/atoms/Button/ReSignButton/ReSignButton.module.scss
@@ -18,4 +18,8 @@
     color: $red;
     border: 1px solid $red;
   }
+
+  &:disabled {
+    opacity: 0.3;
+  }
 }
