+++ src/components/atoms/Button/ReSignButton/index.tsx
@@ -1,6 +1,6 @@
 import styles from '@/components/atoms/Button/ReSignButton/ReSignButton.module.scss';
+import useResignButton from '@/hooks/useResignButton';
 import classNames from 'classnames/bind';
-import { useRouter } from 'next/navigation';
 import { ReactNode } from 'react';
 
 const cn = classNames.bind(styles);
@@ -11,20 +11,15 @@ interface ResignButtonProps {
 }
 
 function ResignButton({ type, children }: ResignButtonProps) {
-  const router = useRouter();
-  const handleCancelClick = () => {
-    // TODO: 이 버튼 누르면 피드 화면으로 이동하도록
-    router.back();
-  };
-  // TODO: delete 클릭은 닉네임과 일치할 때 활성화 되도록
-  const handleDeleteClick = () => {
-    router.push('/');
-  };
+  const { handleCancelClick, handleDeleteClick, isCheckName } =
+    useResignButton();
+
   return (
     <button
       type="button"
       className={cn('button', type)}
       onClick={type === 'cancel' ? handleCancelClick : handleDeleteClick}
+      disabled={type === 'delete' && isCheckName}
     >
       {children}
     </button>
